feat/#116 : 스터디 추천 화면 구현
화면 UI 및 통신 구현

+++ Android/app/src/main/java/com/example/firstproject/data/model/dto/response/UserSuitableStudyDtoItem.kt
@@ -5,7 +5,7 @@ data class UserSuitableStudyDtoItem(
     val meetingDays: List<String>,
     val memberCount: Int,
     val members: List<Member>,
-    val similarity: Int,
+    val similarity: Double,
     val studyId: String,
     val studyName: String,
     val topic: String
